MH-10124 added check if parent is null

+++ modules/matterhorn-distribution-service-streaming/src/main/java/org/opencastproject/distribution/streaming/StreamingDistributionService.java
@@ -296,7 +296,7 @@ public class StreamingDistributionService extends AbstractJobProducer implements
       // Try to remove the file and - if possible - the parent folder
       FileUtils.forceDelete(elementFile);
       File elementDir = elementFile.getParentFile();
-      if (elementDir.isDirectory() && elementDir.list().length == 0) {
+      if (elementDir != null && elementDir.isDirectory() && elementDir.list().length == 0) {
         FileSupport.delete(elementDir);
       }
       if (mediapackageDir.isDirectory() && mediapackageDir.list().length == 0) {
